refactor(eigendb): renamed vector_io package to index
